requirements.txt
openai<1.0.0
requests
python-dotenv
pytest
pytest-cov
coverage
